// client/src/redux/Authentication/actions/enum.ts
enum ActionTypesEnum  {
  POST_SIGNUP = 'POST_SIGNUP',
  POST_SIGNUP_SUCCESS = 'POST_SIGNUP_SUCCESS',
  POST_SIGNUP_FAILURE = 'POST_SIGNUP_FAILURE',
  POST_LOGIN = 'POST_LOGIN',
  POS_LOGIN_SUCCESS = 'POST_LOGIN_SUCCESS',
  POST_LOGIN_FAILURE = 'POST_LOGIN_FAILURE',
  POST_LOGOUT = 'POST_LOGOUT',
  CLOSE_MESSAGE = 'CLOSE_MESSAGE'
}

export default ActionTypesEnum